transform api.translation.v2.function.definition into separate modules grouped by method
